Bump gradle, kotlin and refactor libserialport usage for mobile/desktop.
