+++ android/settings.gradle
@@ -23,7 +23,7 @@ pluginManagement {
 
 plugins {
     id "dev.flutter.flutter-plugin-loader" version "1.0.0"
-    id "com.android.application" version '8.3.2' apply false
+    id "com.android.application" version '8.6.0' apply false
 }
 
 include ":app"
